Made final changes to part 4.

+++ osa4/blogilista/controllers/blogs.js
@@ -1,35 +1,40 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 
-
-blogsRouter.get('/', (request, response) => {
-    Blog
-        .find({})
-        .then(blogs => {
-            response.json(blogs)
-        })
+blogsRouter.get('/', async (request, response) => {
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+    response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response) => {
-    const blog = new Blog(request.body)
-    blog
-        .save()
-        .then(result => {
-            response.status(201).json(result)
-        })
-        .catch(error => {
-            response.status(400).send(error)
-        })
+blogsRouter.post('/', async (request, response, next) => {
+    const body = request.body
+    const user = request.user
+
+    const blog = new Blog({
+        title: body.title,
+        url: body.url,
+        likes: body.likes,
+        author: body.author,
+        user: user._id
+    })
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
-    } catch (exception) {
-        next(exception)
+    const blog = await Blog.findById(request.params.id)
+    const user = request.user
+    if (blog.user.toString() !== user.id) {
+        return response.status(401).json({ error: 'user is not authorized' })
     }
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
